+++ sopa/io/explorer/images.py
@@ -13,7 +13,7 @@ from spatialdata.transformations import Affine, set_transformation
 from tqdm import tqdm
 from xarray import DataArray
 
-from ..._sdata import get_intrinsic_cs, get_spatial_image
+from ..._sdata import add_spatial_element, get_intrinsic_cs, get_spatial_image
 from ...utils.image import resize_numpy, scale_dtype
 from ._constants import ExplorerConstants, FileNames, image_metadata
 from .utils import explorer_file_path
@@ -230,7 +230,4 @@ def align(
     set_transformation(image, {pixel_cs: to_pixel}, set_all=True)
 
     log.info(f"Adding image {image_name}:\n{image}")
-    sdata.images[image_name] = image
-
-    if sdata.is_backed():
-        sdata.write_element(image_name, overwrite=overwrite)
+    add_spatial_element(sdata, image_name, image, overwrite=overwrite)