+++ sopa/patches/__init__.py
@@ -1,1 +1,1 @@
-from .patches import Patches2D, _get_cell_id
+from .patches import Patches2D, _get_cell_id, make_image_patches, make_transcript_patches